Refactor sequences + update notebooks
