+++ src/seqme/__init__.py
@@ -7,9 +7,13 @@ from seqme.core import (
     combine,
     evaluate,
     plot_series,
+    random_subset,
+    read_fasta,
     rename,
     show,
+    shuffle_characters,
     sort,
+    to_fasta,
     to_latex,
     top_k,
 )
@@ -22,12 +26,16 @@ __all__ = [
     "combine",
     "evaluate",
     "plot_series",
+    "random_subset",
+    "read_fasta",
     "rename",
     "show",
+    "shuffle_characters",
     "to_latex",
     "metrics",
     "models",
     "utils",
+    "to_fasta",
     "sort",
     "top_k",
 ]
