+++ src/seqme/core.py
@@ -1,4 +1,5 @@
 import abc
+import random
 from collections import Counter, defaultdict
 from collections.abc import Callable
 from dataclasses import dataclass
@@ -705,7 +706,7 @@ def barplot(
     metric: str,
     *,
     show_deviation: bool = True,
-    color: str = "#14c299",
+    color: str = "#68d6bc",
     x_ticks_label_rotation: float = 45,
     ylim: tuple[float, float] | None = None,
     figsize: tuple[int, int] = (4, 3),
@@ -1006,3 +1007,86 @@ def _get_top_indices(df: pd.DataFrame, metric: str) -> tuple[set[int], set[int]]
         raise ValueError(f"Unknown objective '{objective}' for metric '{metric}'.")
 
     return top_indices_helper(best_cells)
+
+
+def shuffle_characters(sequences: list[str], seed: int = 0) -> list[str]:
+    """
+    Randomly shuffle characters within each sequence, preserving reproducibility.
+
+    Args:
+        sequences: List of input strings to shuffle.
+        seed: Seed for the random number generator to ensure determinism.
+
+    Returns:
+        A new list where each sequence's characters have been shuffled.
+    """
+    rng = random.Random(seed)
+    shuffled = []
+    for seq in sequences:
+        chars = list(seq)
+        rng.shuffle(chars)
+        shuffled.append("".join(chars))
+    return shuffled
+
+
+def random_subset(sequences: list[str], n_samples: int, seed: int = 0) -> list[str]:
+    """
+    Select a random subset of unique sequences with deterministic behavior.
+
+    Args:
+        sequences: The list of input sequences to sample from.
+        n_samples: The number of sequences to sample.
+        seed: The random seed for reproducibility.
+
+    Returns:
+        A list of `n_samples` randomly chosen, unique sequences.
+
+    Raises:
+        ValueError: If `n_samples` exceeds the number of available sequences.
+    """
+    if n_samples > len(sequences):
+        raise ValueError(f"Cannot sample {n_samples} sequences from a list of length {len(sequences)}.")
+
+    rng = random.Random(seed)
+    return rng.sample(sequences, n_samples)
+
+
+def read_fasta(path: str) -> list[str]:
+    """Retrieve sequences from a fasta file."""
+    sequences: list[str] = []
+    current_seq: list[str] = []
+
+    with open(path) as file:
+        for line in file:
+            line = line.strip()
+            if not line:
+                continue  # skip empty lines
+            if line.startswith(">"):
+                if current_seq:
+                    sequence = "".join(current_seq)
+                    if sequence:
+                        sequences.append(sequence)
+                    current_seq = []
+            else:
+                current_seq.append(line)
+
+        # Add the last sequence if present
+        if current_seq:
+            sequence = "".join(current_seq)
+            if sequence:
+                sequences.append(sequence)
+
+    return sequences
+
+
+def to_fasta(path: str, sequences: list[str], headers: list[str] | None = None):
+    """Write sequences to a fasta file."""
+    with open(path, "w") as f:
+        for i, seq in enumerate(sequences):
+            header = headers[i] if headers else f">seq_{i + 1}"
+
+            if not header.startswith(">"):
+                raise ValueError("Missing prefix '>' in header.")
+
+            f.write(f"{header}\n")
+            f.write(f"{seq}\n")
