+++ src/seqme/utils/__init__.py
@@ -5,7 +5,6 @@ from seqme.utils.diagnostics import (
 )
 from seqme.utils.plots import plot_embeddings, plot_hist, plot_kde, plot_violin
 from seqme.utils.projection import pca, tsne, umap
-from seqme.utils.sequences import random_subset, read_fasta, shuffle_characters, to_fasta
 
 __all__ = [
     "feature_alignment_score",
@@ -18,8 +17,4 @@ __all__ = [
     "pca",
     "tsne",
     "umap",
-    "random_subset",
-    "read_fasta",
-    "shuffle_characters",
-    "to_fasta",
 ]
